refactor(table,player): tidy the leftovers from the pill review
Five small things, none of which change what renders:

- `playerTopPillStyle` moves out of `SeatPanel` into its own module. It is
  the geometry four pills share, so a component was an odd place for
  `StatusBar` to have to import it from.
- Drop `justifyContent: space-between` from the table status bar. The
  badge's own `marginLeft: auto` is what pins it right, and it has to be:
  the room-code pill on the left comes and goes, and a lone badge under
  `space-between` alone would sit hard left. A comment now says so.
- Hoist the two `TableControls` placement variants to module scope, keyed
  by placement, instead of rebuilding constants on every render.
- Invert the `controls` ternary in `App` to lead with the condition.
- Scope the shared-pill-height test to the three elements it means, rather
  than counting `height:30px` across the whole document — a count that
  could as easily have been satisfied by an unrelated element, and that
  named no element when it failed.
